+++ CHANGELOG.md
@@ -14,5 +14,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Create draft Professions in the database, rather than using session storage when adding a new profession
+- Fix bug where submitting invalid top level details on a new Profession caused the page to hang
 
 [unreleased]: https://github.com/UKGovernmentBEIS/regulated-professions-register/compare/release...HEAD